added the strains and separete props in the search and results
